validation completed

+++ trello-clone/blueprints/cards_bp.py
@@ -1,87 +1,79 @@
 from flask import Blueprint, request, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from setup import db
 from models.card import CardSchema, Card
-from auth import admin_required
-from datetime import datetime
+from auth import authorize
+from blueprints.comments_bp import comments_bp
 
-cards_bp = Blueprint('cards', __name__, url_prefix='/cards')
+cards_bp = Blueprint("cards", __name__, url_prefix="/cards")
 
-@cards_bp.route('/')
+# Get all cards
+@cards_bp.route("/")
 @jwt_required()
 def all_cards():
-    admin_required()
-    # select * from cards in stmt variable to ;
-    # stmt = db.select(Card).where(db.or_(Card.status != 'Done', Card.id > 2)).order_by(Card.title.desc()) #create statement sql
-    stmt = db.select(Card)
-    cards = db.session.scalars(stmt).all() #execute method, scalars to have singular objects/model instances
-    return CardSchema(many=True).dump(cards) #flask takes care of serialisation
+    # select * from cards;
+    stmt = db.select(
+        Card
+    )  # .where(db.or_(Card.status != "Done", Card.id > 2)).order_by(Card.title.desc())
+    cards = db.session.scalars(stmt).all()
+    return CardSchema(many=True, exclude=["user.cards"]).dump(cards)
 
-# @cards_bp.route("/<int:id>")
-# @jwt_required
-# def one_card(id):
-#     stmt = db.select(Card).filter_by(id=id) # if .where(card.id == id)
-#     card = db.session.scalar(stmt) #execute method, scalars to have singular objects/model instances
-#     if card:
-#         return CardSchema().dump(card)
-    
+# Get one card
 @cards_bp.route("/<int:id>")
 @jwt_required()
 def one_card(id):
-    stmt = db.select(Card).filter_by(id=id)
+    stmt = db.select(Card).filter_by(id=id) # .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
         return CardSchema().dump(card)
     else:
-        # Handle the case where the card with the specified ID is not found
         return {"error": "Card not found"}, 404
 
-#create a new Card
+# Create a new card
 @cards_bp.route("/", methods=["POST"])
-# @jwt_required()
+@jwt_required()
 def create_card():
-    card_info = CardSchema(exclude=['id', 'date_created']).load(request.json)
+    card_info = CardSchema(exclude=["id", "date_created"]).load(request.json)
     card = Card(
-        title=card_info['title'],
-        description=card_info.get('description', ''), #get sets default after
-        status=card_info.get('status', 'To do')
+        title = card_info["title"],
+        description = card_info.get("description", ""),
+        status = card_info.get("status", "To Do"),
+        user_id = get_jwt_identity()
     )
     db.session.add(card)
     db.session.commit()
     return CardSchema().dump(card), 201
 
-
-@cards_bp.route("/<int:id>", methods=["PUT","PATCH"])
-# @jwt_required()
+# Update a card
+@cards_bp.route("/<int:id>", methods=["PUT", "PATCH"])
+@jwt_required()
 def update_card(id):
-    card_info = CardSchema(exclude=['id', 'date_created']).load(request.json)
-    stmt = db.select(Card).filter_by(id=id)
+    card_info = CardSchema(exclude=["id", "date_created"]).load(request.json)
+    stmt = db.select(Card).filter_by(id=id) # .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
-        card.title = card_info.get('title', card.title) #remains unchanged if 'title' is not specified
-        card.description = card_info.get('description', card.description)
-        card.status = card_info.get('status', card.status)
-        db.session.add(card)
+        authorize(card.user_id)
+        card.title = card_info.get("title", card.title)
+        card.description = card_info.get("description", card.description)
+        card.status = card_info.get("status", card.status)
         db.session.commit()
         return CardSchema().dump(card)
     else:
-        # Handle the case where the card with the specified ID is not found
         return {"error": "Card not found"}, 404
 
+# Delete a card
 @cards_bp.route("/<int:id>", methods=["DELETE"])
-# @jwt_required()
+@jwt_required()
 def delete_card(id):
-    admin_required()
-    stmt = db.select(Card).filter_by(id=id)
+    # authorize()
+    stmt = db.select(Card).filter_by(id=id) # .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
+        authorize(card.user_id)
         db.session.delete(card)
         db.session.commit()
-        return {"success": "Card deleted"}
+        return {}, 200
     else:
-        # Handle the case where the card with the specified ID is not found
         return {"error": "Card not found"}, 404
-
     
-
-
+cards_bp.register_blueprint(comments_bp)
